Fly works with PlatformDeployer in separate module.
